+++ src/Rewriter/RewriterFactory.php
@@ -3,15 +3,21 @@
 namespace byrokrat\banking\Rewriter;
 
 /**
- * Create standard set of rewrites
+ * Create standard sets of preprocessors and rewrites
  */
 class RewriterFactory
 {
+    public function createPreprocessors()
+    {
+        return [
+            new NonDigitRemovingRewriter,
+            new LeftTrimRewriter
+        ];
+    }
+
     public function createRewrites()
     {
         return [
-            new LeftTrimRewriter,
-            new ClearingSeparatorRewriter,
             new SwedbankCheckDigitRewriter
         ];
     }